+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `RevitFilterExtensionTest` tests for `ElementFilterExtension`.
 - Add `RevitElementIdExtensionTest` and `RevitElementIdToElementExtensionTest` tests for `ElementIdExtension`.
 - Add `RevitFilterCategoryExtension` tests for `FilterBuiltInCategoryExtension`.
+- Add `RevitSelectElementExtensionTest` tests for select and get elements with filters.
 ### Shapes
 - Add `CreateSwept` to `ShapeCreator` to create a swept solid.
 ### Generator